backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
asgiref==3.9.1
asttokens==2.4.1
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
comm==0.2.0
cryptography==45.0.5
debugpy==1.8.0
decorator==5.1.1
dj-database-url==2.2.0
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
executing==2.0.1
frozenlist==1.7.0
gunicorn==22.0.0
http_ece==1.2.1
idna==3.10
ipykernel==6.27.1
ipython==8.18.1
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
matplotlib-inline==0.1.6
multidict==6.6.3
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
pillow==11.3.0
platformdirs==4.1.0
prompt-toolkit==3.0.41
propcache==0.3.2
psutil==5.9.6
psycopg2-binary==2.9.10
pure-eval==0.2.2
py-vapid==1.9.2
pycparser==2.22
Pygments==2.17.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pywebpush==1.14.0
pywin32==306
pyzmq==25.1.2
requests==2.31.0
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
tornado==6.4
traitlets==5.14.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.12
whitenoise==6.6.0
yarl==1.20.1
